Add `regexp/sort-character-class-elements` rule that same `regexp/order-in-character-class` rule
